add preview wide layout
