+++ src/components/color/detail.tsx
@@ -1,10 +1,18 @@
 "use client";
 
+import type { OklchColor } from "~/lib/types";
 import { useColor, useMode } from "~/app/store";
 import { useColorQuery } from "~/app/query";
-import { createColor } from "~/lib/color";
+import { useMemo, useCallback } from "react";
+import {
+  createColor,
+  measureColor,
+  contrastColor,
+  switchColorCss,
+} from "~/lib/color";
 import ColorPicker from "./picker";
-import ColorAnalysis from "./analysis";
+import ColorContrast from "./contrast";
+import Progress from "../ui/progress";
 import Wrapper from "../ui/wrapper";
 import Link from "next/link";
 
@@ -27,37 +35,55 @@ export default function ColorDetail() {
   const modeOklab = mode === "oklab";
   const modeOklch = mode === "oklch";
 
-  const colorMode = modeRgb
-    ? rgb.css
-    : modeHsl
-      ? hsl.css
-      : modeHwb
-        ? hwb.css
-        : modeLab
-          ? lab.css
-          : modeLch
-            ? lch.css
-            : modeOklab
-              ? oklab.css
-              : modeOklch
-                ? oklch.css
-                : hex;
-
+  const colorMode = switchColorCss(mode, color);
   const colorPicker = colorQuery ? false : true;
+
+  const { brightness, luminance } = measureColor(rgb.color);
+
+  const { white, black } = useMemo(
+    () => ({
+      white: contrastColor(rgb.color, { r: 1, g: 1, b: 1 }),
+      black: contrastColor(rgb.color, { r: 0, g: 0, b: 0 }),
+    }),
+    [rgb],
+  );
+
+  const previewColor = useCallback(
+    (newColor: Partial<OklchColor>) =>
+      switchColorCss(
+        mode,
+        createColor({ mode: "oklch", ...oklch.color, ...newColor }),
+      ),
+    [mode, oklch],
+  );
+
+  const trackBrightnessLeft = previewColor({ l: 0.2, c: 0 });
+  const trackBrightnessRight = previewColor({ l: 1, c: 0 });
+
+  const trackLuminanceBlue = previewColor({ h: 250 });
+  const trackLuminanceGreen = previewColor({ h: 180 });
+  const trackLuminanceYellow = previewColor({ h: 90 });
+  const trackLuminanceWhite = previewColor({ c: 0 });
 
   return (
     <Wrapper
       as="article"
       aria-labelledby="color"
-      maxWidth="1024"
-      className="grid gap-y-4"
+      maxWidth="1280"
+      className="grid gap-y-8"
       style={{ ["--currentColor" as string]: colorMode }}
     >
-      <header className="inline-grid gap-y-3">
+      <h1 id="color" className="sr-only">
+        {colorMode}
+      </h1>
+      <section
+        aria-label="details"
+        className="mx-auto inline-grid w-full max-w-5xl gap-y-4"
+      >
         <Link
           aria-label={`view color ${hex}`}
           href={`/color/${hex.replace("#", "")}`}
-          className="frame inline-grid w-full rounded-lg focus-visible:z-69"
+          className="frame rounded-lg focus-visible:z-69"
           prefetch={false}
         >
           <span
@@ -65,69 +91,88 @@ export default function ColorDetail() {
             style={{ backgroundColor: "var(--currentColor)" }}
           ></span>
         </Link>
-      </header>
-      <section aria-label="legacy">
-        <p>
-          <code>{hex}</code>
-        </p>
-        <p
-          className={
-            modeRgb ? "text-neutral-800 dark:text-neutral-200" : undefined
-          }
-        >
-          <code>{rgb.css}</code>
-        </p>
-        <p
-          className={
-            modeHsl ? "text-neutral-800 dark:text-neutral-200" : undefined
-          }
-        >
-          <code>{hsl.css}</code>
-        </p>
-        <p
-          className={
-            modeHwb ? "text-neutral-800 dark:text-neutral-200" : undefined
-          }
-        >
-          <code>{hwb.css}</code>
-        </p>
-      </section>
-      <section aria-label="modern">
-        <p
-          className={
-            modeLch ? "text-neutral-800 dark:text-neutral-200" : undefined
-          }
-        >
-          <code>{lch.css}</code>
-        </p>
-        <p
-          className={
-            modeOklch ? "text-neutral-800 dark:text-neutral-200" : undefined
-          }
-        >
-          <code>{oklch.css}</code>
-        </p>
-        <p
-          className={
-            modeLab ? "text-neutral-800 dark:text-neutral-200" : undefined
-          }
-        >
-          <code>{lab.css}</code>
-        </p>
-        <p
-          className={
-            modeOklab ? "text-neutral-800 dark:text-neutral-200" : undefined
-          }
-        >
-          <code>{oklab.css}</code>
-        </p>
+        <div>
+          <p>
+            <code>{hex}</code>
+          </p>
+          <p
+            className={
+              modeRgb ? "text-neutral-800 dark:text-neutral-200" : undefined
+            }
+          >
+            <code>{rgb.css}</code>
+          </p>
+          <p
+            className={
+              modeHsl ? "text-neutral-800 dark:text-neutral-200" : undefined
+            }
+          >
+            <code>{hsl.css}</code>
+          </p>
+          <p
+            className={
+              modeHwb ? "text-neutral-800 dark:text-neutral-200" : undefined
+            }
+          >
+            <code>{hwb.css}</code>
+          </p>
+        </div>
+        <div>
+          <p
+            className={
+              modeLch ? "text-neutral-800 dark:text-neutral-200" : undefined
+            }
+          >
+            <code>{lch.css}</code>
+          </p>
+          <p
+            className={
+              modeOklch ? "text-neutral-800 dark:text-neutral-200" : undefined
+            }
+          >
+            <code>{oklch.css}</code>
+          </p>
+          <p
+            className={
+              modeLab ? "text-neutral-800 dark:text-neutral-200" : undefined
+            }
+          >
+            <code>{lab.css}</code>
+          </p>
+          <p
+            className={
+              modeOklab ? "text-neutral-800 dark:text-neutral-200" : undefined
+            }
+          >
+            <code>{oklab.css}</code>
+          </p>
+        </div>
       </section>
       {colorPicker ? (
-        <div className="my-4 grid border-y border-y-neutral-400 py-8 dark:border-y-neutral-700">
+        <div className="grid border-y border-y-neutral-400 py-8 dark:border-y-neutral-700">
           <ColorPicker showTextbox={false} />
         </div>
       ) : null}
-      <ColorAnalysis color={color} mode={mode} />
+      <div className="inline-grid gap-8 md:grid-cols-2">
+        <div className="grid gap-y-3">
+          <p>Brightness</p>
+          <Progress
+            value={brightness}
+            style={`linear-gradient(to right, ${trackBrightnessLeft}, ${trackBrightnessRight})`}
+          />
+        </div>
+        <div className="grid gap-y-3">
+          <p>Luminance</p>
+          <Progress
+            value={luminance}
+            style={`linear-gradient(to right, ${trackLuminanceBlue}, ${trackLuminanceGreen} 25%, ${trackLuminanceYellow} 50%, ${trackLuminanceWhite} 80%)`}
+          />
+        </div>
+      </div>
+      <div className="mt-4 inline-grid gap-x-8 gap-y-12 md:grid-cols-2">
+        <ColorContrast reflect={true} color={white} />
+        <ColorContrast reflect={false} color={black} />
+      </div>
     </Wrapper>
   );
 }
